Convert defense parameters and GUI inputs to floating point
- Support floating-point armor, shields, and point_defense values across ComponentConfig, template I/O, and component instantiation
- Implement continuous damage mitigation via random.uniform in Defenses.calculate_mitigation, rounding only the final total
- Update Unit Designer and Retrofit Wizard entry fields to format float parameters using :g instead of integer truncation
- Parse floating-point inputs in defense reader helpers and preserve precision in design summary views
- Add regression test coverage for float defense validation, GUI parsing/sync, and combat mitigation

Validation: 66 builtin templates checked with 0 errors; 391 targeted unit, component, editor, and validation tests passed.

+++ custom_unit_templates.py
@@ -209,7 +209,7 @@ def calc_weapons_hull_cost(turrets: List['TurretConfig']) -> float:
     return Weapons.calc_hull_cost(turrets)
 
 
-def calc_defenses_hull_cost(armor: int, shields: int, point_defense: int) -> float:
+def calc_defenses_hull_cost(armor: float, shields: float, point_defense: float) -> float:
     """Compute the hull cost of a Defenses component from its stats."""
     return Defenses.calc_hull_cost(armor, shields, point_defense)
 
@@ -354,9 +354,9 @@ class ComponentConfig:
 
     # Defenses
     has_defenses: bool = False
-    armor: int = 0
-    shields: int = 0
-    point_defense: int = 0
+    armor: float = 0.0
+    shields: float = 0.0
+    point_defense: float = 0.0
     # hull cost is computed: see defenses_hull_cost property
 
     # Constructor
@@ -914,9 +914,9 @@ def template_from_dict(key: str, d: Dict[str, Any]) -> CustomUnitTemplate:
         turrets=turrets,
 
         has_defenses=d.get("has_defenses", False),
-        armor=d.get("armor", 0),
-        shields=d.get("shields", 0),
-        point_defense=d.get("point_defense", 0),
+        armor=float(d.get("armor", 0.0)),
+        shields=float(d.get("shields", 0.0)),
+        point_defense=float(d.get("point_defense", 0.0)),
 
         # --- Fixed-cost components ---
         has_constructor_component=d.get("has_constructor_component", False),
